More comments

+++ csv.js
@@ -1,10 +1,12 @@
-const fs = require('fs');
 const createCsvWriter = require('csv-writer').createObjectCsvWriter;
+const fs = require('fs');
+
 
 // New date formatted to use as file name.
 const date = new Date().toISOString().split('T')[0];
 
-// csvWriter instance defines headers, sets path using date above
+
+// Create csvWriter
 const csvWriter = createCsvWriter({
     path: `/Volumes/E-Drive/Programming/github/THProjects/scraper/data/${date}.csv`,
     header: [
@@ -16,20 +18,20 @@ const csvWriter = createCsvWriter({
     ]
 });
 
+// Write function
 function write(records) {
   const dataFolder = '/Volumes/E-Drive/Programming/github/THProjects/scraper/data';
   const todaysFile = `/Volumes/E-Drive/Programming/github/THProjects/scraper/data/${date}.csv`;
 
   // Check for data folder, create it if it doesn't exist
-  if (!fs.existsSync(dataFolder)) {
-    fs.mkdirSync(dataFolder);
-  }
+  if (!fs.existsSync(dataFolder)) {fs.mkdirSync(dataFolder)}
   // Check for todays file, delete it if it does exist
-  if (fs.existsSync(todaysFile)) {
-    fs.unlinkSync(todaysFile);
-  }
+  if (fs.existsSync(todaysFile)) {fs.unlinkSync(todaysFile)}
+  
   // Write the csv file to the data directory
-  csvWriter.writeRecords(records).then(() => console.log(`Success!\nToday's data is located at \n${todaysFile}`)).catch((error) => console.log(error));
+  csvWriter.writeRecords(records)
+    .then( () => console.log(`Write successful, file is located at:\n${todaysFile}`) )
+    .catch( (error) => console.log(error) );
 }
 
 module.exports.write = write;
